[FAB-10113] test idemix MSP in the e2e
This change set introduces support to test the client idemix signer and the
peer idemix verifier in the e2e test. This is accomplished by adding a third
organization with MSP of type idemix, and adding an extra chaincode query
where the creator signs using idemix. Notice that no extra peer is
introduced, since peers cannot operate as idemix endorsers yet with a local
MSP of type idemix.

Change-Id: Ice078d75f12e901ba580583791a8574a14b96802
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>