[FAB-10151] reorder sections in configtx.yaml
Most of the configtx.yaml documents in the tree contained invalid YAML.
This wasn't detected before now because bugs in go-yaml have hidden the
issue. These changes need to be made before the parser is bumped to help
ensure things don't break.

A draft release note has been added to the tree to highlight the issue
as we approach the end of our current release cycle.

Change-Id: I8d9cd9e2278ccd51ffd136fd17e3d858f6b56b2d
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>