[FAB-10173] Done in testdata sample validation plugin
The validation test has a sample implementation of a validation plugin.
There is a missing call to Done() in the code.
Since this is just a test, it doesn't affect production runtime,
but it might still serve as a reference of how to write a validation plugin.

Therefore, to set a good example - we should also add a call to Done()
on the state object after we obtain it.

Change-Id: Ie96a8fdf583aa2c8c64ceb119810bb1764ba3041
Signed-off-by: yacovm <yacovm@il.ibm.com>