[FAB-10201] handle duplicate tx detection errors
GetTransactionByID is used to detect duplicate txids in the ledger. This
function can return an error if there was an I/O problem (and so the duplicate
status of the tx couldn't be properly judged). In that case we must
abort processing on that chain, lest we suffer a state fork.

Change-Id: I32a8ea9c0803f7a38b4c6eb3ea30b4ce1a651456
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>