FAB-10826 Change ValidateSpec to ValidatePath
This is part of a CR series to remove the explicit dependency of the
chaincode spec and deployment spec from the platforms package.
Ultimately, these messages were designed for fabric v0.5/v0.6 and would
not have been recycled given more time.  This is a step in allowing us
to use a different, more fabric v1.0 natural format for describing
chaincodes.

Change-Id: Ibbd20f3ce85f226e4c2cc03b3e2fd70081815ba6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>