FAB-10833 Fix ccmetadata package structure
The ccmetadata is imported by utils and by other platforms, but is also
part of the platforms interface definitions itself.  This CR moves the
interface definition into the platforms package where it belongs, and
also updates the implementation to accept an interface instead of
requiring the chaincode deployment spec.  It is part of a series to
remove the deployment spec from the platforms package.

Change-Id: I7ab07c98a88fdf67129c29320ac1433e8b466695
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>