[FAB-10240] Remove duplicated checking on TxID
When doing endorsement, the TxID is checked in
core/common/validation/msgvalidation.go to make sure it equals to
SHA256(nonce+creator).

After that, check whether it is empty again in the core/endorser/endorser.go.
However, the later checking actually won't happen because of the previous validation.

This patchset removed the unused empty checking code.

Change-Id: I99de8e833432200196741b55555d9006f8cb86a0
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>