FAB-10827 Remove meaningless file write from test
The node platform test currently creates a dummy file, then stats it and
reads it in, then creates a tar header from the stat result, then
modifies the tar header.  All of this is unnecessary code when the test
can simply create the tar custom tar header directly.  This CR does just
that.

Change-Id: I9ea0ada2df5441b102faf7ead3ecc9457c7b3d93
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>