FAB-10283 Make configtxlator handle maps correctly
The current marshaling scheme employed in configtxlator is the
protolator package, which converts JSON back into protobuf.  Because
updates are computed on the binary protobuf data, we need for the same
configtxlator server to deterministically marshal the same message the
same way.

This CR adds a 'MostlyDeterministicMarshal' function to the protolator
package so that marshaling within a single configtxlator server or
process will be consistent.  As noted in the comment, this should not be
misconstrued as true deterministic marshaling, and only works because
configtxlator is a single binary which interoperates only with itself.

Change-Id: I1b942a9fba9e7ea0b94fdfbdd1cac595978994b7
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>