FAB-10303 Clean up some assorted struct init
Many of our tests and some of our code are still using order based
initialization for structs.  This is very brittle and against bad
practices.  In an attempt to upgrade the version of protobuf to v1.1.0,
all of files in this changeset broke, because the number of struct
members in the generated code changed.  Although this does not include
any of the proto updates itself, this CR serves as a prerequisite to
doing so.

Change-Id: Ib09ac38735b24894bd413ddaaf59288678c6151d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>