FAB-10331 Fix mockery mock generation
The comments in the go files which are supposed to generate the mockery
mocks are malformed.  They use:

 // go:generate ...

instead of

 //go:generate

They also assume they are being executed from the root directory, rather
than the current directory.  This CR modifies those statements to
correct those errors, regenerates all the mocks, and updates the license
check to ignore mockery style mocks.

Change-Id: I7d8ca47254459e2929286cf614e09893b254d35d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>