[FAB-10462] Discovery: use ChannelWriters, not Readers
The discovery access control checks for channel readers,
however - if the channel admins want to enforce that only
admins or peers can pull blocks from the orderer/peers,
they would set the reader policy to be OR(admin, peer)
and that blocks every client from using service discovery.

It would make sense to change the policy that is checked to be
channel application writers instead of readers, since if you
should be allowed to use an endorsement (write to the channel)
you should also be allowed to see who are the peers that you should endorse to.

Change-Id: I6e54de2017339060f3f14e3289138fe51d479bec
Signed-off-by: yacovm <yacovm@il.ibm.com>