FAB-10976 Move lifecycle to its own package
Because of some errors with cyclical imports for the chaincode package
tests which are not written as _test packages, we need to define a
structure to represent a chaincode which is not the
ChaincodeDeploymentSpec but which is also not inside of the chaincode
package.  Defining this in the lifecycle portion makes sense, so as a
prerequisite, moving lifecycle from the chaincode package, to a
subpackage called lifecycle.

Note: This introduces some stuttering, that will be remedied in a later
CR.

Change-Id: I8883b9fc39d7b23d42ba6df31389e5a9df248457
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>