FAB-11038 Disable chaincode lookup cache swizzling
Per comments left in the ccprovider package, the chaincode cache was
always intended to be on by default (in fact, in production, it is
always enabled).  However, the option to toggle it on and off was
preserved for some poorly written tests.  Since there seem to be no
tests which fail when the cache is enabled, removing this option
entirely.

Change-Id: Id2a74c538e9e3c72b2bafd9bb324792da2040bb4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>