[FAB-10505] Memory coherent channel.Resources fetching
The discovery calls peer.GetChannelConfig to fetch the latest
channelconfig.Resources object and extract from it the the
Sequence() of the channel (the config block sequence).

However, it might read a stale reference of the channelconfig.Resources
because that reference is updated via an atomic pointer,
while the peer.GetChannelConfig doesn't call an atomic load, but
relies on the channels map's RLock which the callback that
updates the map - doesn't lock.

This can be easily solved via just calling GetStableChannelConfig
instead of GetChannelConfig

Change-Id: I7374ebdc8cb991d0c38a1d39e7f39c0e717dce63
Signed-off-by: yacovm <yacovm@il.ibm.com>