FAB-10982 Add direct accessor for LSCC to get CDS
In many cases, there is no need to access the LSCC state as a chaincode.
In these cases, it's much easier to simply pass a reference to the LSCC
and query it directly.  This CR adds a new method to get the chaincode
deployment spec for a chaincode as well as a new ginkgo style unit test
for this function.

Change-Id: Iad5480d2e1485f5a42f9cb1f253c0ca9b651ff25
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>