FAB-10984 Prepare to move SCC attributes into SCC
The scc package currently enumerates all of the system chaincodes, their
names, versions, etc. and instantiates them.  This is the wrong pattern
to follow, as system chaincodes can describe themselves, and
the wiring code which assembles a working peer should simply pick the
system chaincodes it would like to include.  This is a first step
towards removing the built-in system chaincodes from the scc package and
pushing it into peer/node/start.go

Change-Id: Ibb249f052e680f989f92d3f94b7f1f1003e194ca
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>