FAB-11000 Remove unnecessary context
There is a context passed into all of the Start/Stop methods for the
chaincode container management.  Historically, the only ever reference
to it was via a value on the context to get back the chaincode support.
However, since this dependency has been remeoved, it no longer serves
any point, other than to complicate method signatures and cause general
confusion.

It's clear whenever it was wired, the intent was to eventually make the
dockerclient etc. have cancelable contexts.  However, this work was
never done, and it is sufficiently low priority as to not even appear in
the backlog.  It is trivial enough to add back the context dependency if
it's ever shown that it's useful.

Change-Id: I8bcf0922594a5e40cb187687849b49ad5653ae89
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>