FAB-11030 Remove skipped tests from chaincode
The chaincode package has many many lines of tests which are currently
skipped.  Most of these tests do not even work anymore, and they are
maintained only to the extent that they still compile.  This CR simply
deletes all of the tests and helper functions which are not used
anymore, as they are a burden on development.

Change-Id: Ib61d2db34b8335c1c3463e223eefe12224b51762
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>