FAB-11034 Replace context with explicit params
Presently, there is a "context" which is passed into the chaincode
package.  Rather than actually be used for its intended purpose (of
waiting for cancelation), the context is never actually checked.
Instead, it simply carries the transaction simulators around in an
opaque fashion.

It's much better to make these transaction simulators to be passed in
explicitly.  Additionally, some other facets like TxID and the assorted
proposals really fit much more naturally with the simulators and not
with the chaincode context.  So, this CR brings them in as well.

Change-Id: I7e0c43af22f1e96b8d302e5cf877e7beec36598b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>