FAB-11036 Remove the transaction parms from cccid
The chaincode context structure help the channel id, the chaincode name,
chaincode version, and other things unrelated to identifying a chaincode
like the proposal bytes.  These other fields are also stored in the
transaction parms, so, they are not needed and are confusing when in the
chaincode context.  This CR removes them.

Change-Id: I892bd24cc621800cd1b089d90d0a603e112ae84c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>