[FAB-10141] Fix bug in genesis block persistence
This CR fixes a bug in the kvledger Provider. This bug
prevented to persist a copy of the genesis block with the
ledger provider. Note that This persistence is not used by
the currently offered functionalities. However, this could be
useful for some of the administration related functions that
we may offer in the future versions.

Change-Id: Ia6876016035ac758d7aaccdb383401e78b8e742f
Signed-off-by: manish <manish.sethi@gmail.com>