[FAB-9134] Fix data race in DockerBuild
A data race was detected when running TestGenerateDockerBuild. The race
was due to concurrent read/write of a bytes buffer used to capture the
streaming output of the chaincode build container.

The race can be avoided by closing the CloseWaiter returned by the
attach call and waiting for the stream processing to close before
reading the output.

Change-Id: If34bb11f2a263733147cf2262f25a3617c27e342
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>