[FAB-9150] race in TestRegister_ExpiredIdentity
The event server starts a go routine to distribute events to registered
handlers and the expired identity test reaches into the handler to force
the identity associated with the handler to expire in the past. The
current implementation expires the sessions and publishes a block to
trigger the event processing without serialization and this is reported
as a race.

A mutex was added to serialize reads and updates of the expiration time.
Although this is not strictly needed for the production code, it
resolves the race without refactoring to introduce a fake clock or
similar dependency.

Change-Id: I0cefd847a6e564d81099e116d463bf22231b4fc3
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>