[FAB-10963] Discovery: filtering in membership queries
This change set adds support to the discovery service to filter
channel membership queries according to endorsement capabilities.

Now the client can pass invocation chains to the query, and the
discovery service would only return peers that satisfy the invocation
chain:
1) Peers that have chaincodes installed on them
2) Peers that are authorized by a given collection configuration

This change set also enhances the golang client implementation with
an ability to pass in the Peers() method - an invocation chain.

To preserve backward compatibility as much as possible,
the invocation chain is passed as a variadic parameter.

Change-Id: I28b4bcd88d76a40e5dfcdeba467e1269dbdd0101
Signed-off-by: yacovm <yacovm@il.ibm.com>