[FAB-11047] Enlarge discovery client memoization cache
The discovery server side has a credential cache that reuses
results of access control checks for tuples of requests and signatures
to save CPU cycles.

The idea is that when clients send a request that was sent in the past, they can
skip signinig it and send the signature that was produced in the past in order
for the service on the peer to find the signature and request digest in its cache
and not do costly signature checks.

The discovery client implementation which is used by the go-SDK, caches
only the latest request sent to the service which isn't ideal.

This change set changes this, and introduces a small cache to the
client implementation.

Change-Id: Iade112d2a0cd1fac7dc7968105c652ad7f53d88c
Signed-off-by: yacovm <yacovm@il.ibm.com>