FAB-11306 update shim proto to support pagination
For pagination, chaincode would pass a query metadata with
range/rich queries. The query metadata would contain a page size
and a bookmark.

As a result of a range/rich query with pagination, the peer
should return a query response metadata which would contain
a fetched records count and a bookmark.

Hence, this CR defines two new messages (1) QueryMetadata and
(2) QueryResponseMetadata in the protos_/peer/chaincode_shim.proto

Change-Id: I145681ee5d784e14b2f4386269ddcf2a0812e1ab
Signed-off-by: senthil <cendhu@gmail.com>
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>