[FAB-11172] Token transaction proto message
This changeset contains the definition of a transaction message accommodating
token issue, and transfer, TokenTransaction. For TokenTransaction to be
possible to integrate in the common.Envelope structure recognised by the
orderers, we extend common.HeaderType with a "Token" type. HeaderType
indicates the type of message (payload) included in a common.Envelope.
Peers who receive common.Envelope messages with Envelope.Header including a
HeaderType of "Token" would parse Envelope.Payload as TokenTransaction.

Change-Id: Ica1a5407a14a22f5941de40e43b6ef0c71f83486
Signed-off-by: Elli Androulaki <lli@zurich.ibm.com>
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Mathias Bjoerkqvist <mbj@zurich.ibm.com>