FAB-11508 Add consenter metadata to channel config
This changeset:

1. Adds an opaaue Metadata field to the Channel/Orderer/ConsensusType
value. This is where consensus plugin authors are expected to encode
information related to the consenter nodes. An associated getter has
been added so that consensus implementations can actually retrieve the
metadata at runtime.

2. Allows protolator to work with orderer metadata. Specifically, it
defines a ConsensusTypeMetadataFactory interface and map in
protos/orderer that protolator will consult in order to translate
metadata proto messages to JSON.

Change-Id: I2bfc7e293a370d43e9785df2e3572535106ca8b6
Signed-off-by: Kostas Christidis <kostas@christidis.io>
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>