Ledger-lscc: Introduce queriabilty on to-be state
This CR introduces a query executor that operates on statedb and
one or more state-updates on top. This enables a queriability
on to-be committed state before committing the state.

The combiner expects the input - a slice of queriable kv-data
(either statedb or state-update batches). The slice is assumed to
contain the kv-data in the decending order of updates - i.e,
a kv-data present at an index 'i' is assumed to overwrite the
kv-data present at the index 'i+1'. For instance, if a key is
deleted from the kv-data present in the lowest index, the combiner
will treated the key as deleted.

The combiner offers a simple query interface that provides 'Get'
and 'Range' query options.

This is expected to be used in the state listener so that a state
listener can query the to-be committed state.

FAB-11560 #done

Change-Id: Ie24e1e4e859b0e3393bc0854700a8a6579159167
Signed-off-by: manish <manish.sethi@gmail.com>