[FAB-11586] Raft communication layer, part 2
This change set implements the higher level of the raft
communication layer which is to be used by the consumers
of the communication layer.

It uses the base communication layer from FAB-11585 and
exposes a higher level API over it for sending messages,
via an RPC object, and also implements a gRPC service
that implements the Cluster protobuf API and passes
the messages to the base communication layer.

Change-Id: Id5f5f60179b4cac5ef7ab88bc4c9f5b7425391e6
Signed-off-by: yacovm <yacovm@il.ibm.com>