Declare ledger dependency for coll membership info
This CR enables ledger to take an external dependency that ledger
can use for determining whether the peer is a member of
a given collection on a channel.

The intended usage of this dependency is to detemine a situation
when the peer becomes a memeber of a new collection.

FAB-11436 #done

Change-Id: Id5d342ea12974aa16d84aaa6ae99db5e04716877
Signed-off-by: manish <manish.sethi@gmail.com>
Signed-off-by: nirro <nirro@il.ibm.com>