[FAB-11769] validate cc ns change for meta writes
Currently, transactions are validated by a validation plugin for a namespace
if they write to keys into that namespace. More specifically, if the write
set is not the empty set for that namespace. With state-based endorsement,
this conditions should be changed to ensure that transactions that only modify
metadata in a namespace also undergo validation.

Change-Id: I89009ecb4a08510e9b1f2deb2c2b9b7ec7c9f7bf
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>