[FAB-11802] State-based val. must at least check CCEP
FAB-9473 addressed the following issue: if a transaction doesn't perform
writes in the namespace of a chaincode, that chaincode's EP must still be
checked for the chaincode that is being invoked directly (and not by way of
a cc2cc invocation). This behaviour must not change with the
state-based validator.  However, we cannot adopt the 1.2 behaviour of
always checking the chaincode EP as that would invalidate the premise of
state-based endorsement. The following approach is taken: the state-based
validator will always validate the chaincode endorsement policy in case the
transaction did not require checking any other endorsement policy (notably,
any state-based one).

Change-Id: Ib8c1a1cdbf001994b12db1b8762a96369ab20855
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>