[FAB-11860] KeyLevelEndorsement cap. for 1.3 peers only
The KeyLevelEndorsement capability was mistakenly assigned to 1.2 peers,
while the development of [FAB-8812] wasn't complete yet. Since it is
never used by 1.2 peers, we can safely re-assign it to 1.3 peers only,
and use it to properly gate the behaviour of 1.3 nodes with respect to
key level endorsement.

After this change set is merged, peers will behave as follows:
 -) 1.2 peers: KeyLevelEndorsement will return true if the channel is
    defined with the "V1_2" capability. However, 1.2 peers never call
		this function, and it has been removed by [FAB-11870].
 -) 1.3 peers: KeyLevelEndorsement will return true if the channel is
    defined with the "V1_3" capability.

In conclusion, it is safe to use KeyLevelEndorsement to gate 1.3-specific
behaviour after this change set is merged: 1.2 peers never use it anyway,
and 1.3 peers will possess this capability only if the "V1_3" channel
capability is set.

Change-Id: Iff249cbf129670feae85e6c2b74b8e009c3f0eb7
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>