[FAB-11905] Provide full set of endorsers in filter
The Endorsers function of the Discovery client was
changed to accept a filter that can both filter out unwanted
endorsers and also prioritize them. The filter provides the
complete set of endorsers so that filter implementations
may have a complete picture of the state of the endorsers
(for example, maximum block height of the set of endorsers)
and perform logic based on that information.

Change-Id: Id6b202d41ce42fd539403e59384676f33497cead
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>