Fix historydb issue for keys containing nil bytes
historydb code assumes that the keys does not contain a nil byte.
In the presence of such keys, a historydb query can behave in an
unpredictable way that ranges from including the wrong results
to a panic crash.

The main issue is that when the keys have nil bytes, additional
keys fall in the results of the range query that is formed for
scanning the history of a key. This CR provides a simple fix
in which these non-relevant results will be skipped.

Though, ideally, the keys layout should be such that the non-relevant
results of a scan should be zero, however, that solution requires a
change in key formats and will require a rebuild of the historydb
and hence not suitable for an interim release and will potentially
be provided with 2.0 release.

done #FAB-11244

Change-Id: I843f1babc2673d3a38c47e9e04a660203180a3de
Signed-off-by: manish <manish.sethi@gmail.com>