[FAB-11928] Idemix attributes in chaincode
Added a new function in the chaincode library to allow
the caller to get values of attributes that are exposed
as part of the Idemix identiy. Currently the only attributes
that are available are 'ou' and 'role'.

To make it easier for chaincode writers to not have import extra
depedencies, the cid and attrmgr code was moved down into shim.

Change-Id: Icfa45e213b9a8d963ae689f557347777ca76d9f1
Signed-off-by: Saad Karim <skarim@us.ibm.com>