[FAB-11973] failing test in statebased validation
With [FAB-11948], retrieval of state from a collection may return an error
in case the collection configuration isn't defined. The key-level validation
code must handle this case without failing, given that failure in validation
implies halting the processing on the chain. The chosen approach is to treat
the error as a benign retrieval failure, crucially, one that will
deterministically occur at every committing peer. As such, validation can
continue as if that key didn't have state-based endorsement policies. The
validator may therefore mark the transaction as valid, but the ledger will
mark it as invalid given the attempt to write to an undefined collection.

Change-Id: I2e5aaf0d05ec803873edcb92a3b9377e440b1ca1
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>