[FAB-12155] IssuerValidator Impl
This change-set does the following:
1. Implementes the IssuerValidator interface
in the following way: The Validate  method checks if
the passed credential correspond to that of a valid
member of a Fabric channel the validator is bound to.
The Fabric channel is representetd by an MSP manager that collects
all the MSPs registered in that channel.
2. It modify the manager.Manager to use the
implementation introduced in point 1.
3. Refactor the interfaces to avoid circual package dependecy
and local interfaces proliferation

Change-Id: I0b48b17f3e6246f34be3a1c28842352649fed917
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Mathias Bjoerkqvist <mbj@zurich.ibm.com>