FAB-12229 Return closed err chan until Kafka start
Fixing some race conditions in the orderer code changed the behavior of
the Errored() function to return 'nil' until startup when the previously
racy behavior was to return a closed channel.

This CR changes the return of nil to be a closed channel, and eliminates
the previous setting of the errChan in the newChain function (as it is
ignored until startup completes which overwrites this member).

Change-Id: Ie69b7e710940d9c5fb8dbf8e925cd18a75169fb1
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>