[FAB-12704] Detect If orderer node is in channel
This change set adds support for detecting if the ordering node
is in a given channel.

In orderer/common/cluster, the added method receives a Puller
for the channel, and:
- Pulls the last block from some orderer
- Pulls the last config block from some orderer, based
  on the last config from the block metadata from the last block pulled
- Invokes the passed predicate on the config block.

In the etcdraft consenter code I added an implementation for the predicate
that can be re-used for avoiding creation of new chains when not needed.

Change-Id: I31b0c6f868b0eefa55a1f364253b343d3b3514ea
Signed-off-by: yacovm <yacovm@il.ibm.com>