[FAB-12763] test flake in TestStreamAbort
In rare cases, the gRPC infrastructure returns an internal error
instead of a context canceled error.

This is because the gRPC stack doesn't handle properly sending of messages
that are un-initialized interface references and tries to serialize them.

As a result, the server routine in the test sends a message before the gRPC
infrastructure closes the stream.

This is a race, but not a data race, and the simplest thing is
to just accept both type of errors.

Change-Id: I3cb438ae029039a817474a3eae9552a84374960b
Signed-off-by: yacovm <yacovm@il.ibm.com>