[FAB-12926] Validate hash chain when listing blocks
This change set adds validation of the hash chain to the ChainInspector.

We need to validate the hash chain (backward hash pointers) when iterating over
the system channel when inspecting the channels that exist (which is
what the ChainInspector does).

Even though the block puller that fetches the channels validates the hash chain in batches,
it doesn't verify the hash pointers between the batches themselves.

Since the onboarding puller (unlike the etcdraft one) is equipped with a no-op
signature validator and relies on the validity of the bootstrap block
in conjunction with the integrity of the pulled hash chain,
it doesn't verify the signature of the last block of a pulled batch,
thus we need to add this check into the code of the channel lister.

Change-Id: I8320d093f8d6f5a81291e6377536b0ebdcc83c48
Signed-off-by: yacovm <yacovm@il.ibm.com>